backend/pyproject.toml
[project]
name = "alphastream-dataquest"
version = "0.1.0"
description = "Real-time trading recommendations powered by Pathway streaming RAG"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    # Core framework
    "pathway>=0.8.0",
    # API & async
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "websockets>=12.0",
    # LLM (OpenRouter compatible)
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    # Data processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "nltk>=3.8.0",
    # Financial data
    "yfinance>=0.2.32",
    "ta>=0.10.2",
    # Utilities
    "python-dotenv>=1.0.0",
    "edgartools>=5.10.2",
    "reportlab>=4.4.9",
    "matplotlib>=3.10.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

# GPU support (optional - installs CUDA libs)
gpu = [
    "sentence-transformers>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Force CPU-only torch via index
[tool.uv]
index-url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
